homework_7_streaming/workshop/requirements.txt
# Python dependencies for producer and consumer scripts
# Run: pip install -r requirements.txt

kafka-python==2.0.2
pandas==2.2.2
pyarrow==16.1.0
psycopg2-binary==2.9.9
